esp32/esp32_validated_modules/mqtt_subscriber_publisher/django_web/requirement.txt
asgiref==3.8.1
Django==5.1.5
django-bootstrap5==24.3
gunicorn==23.0.0
mysqlclient==2.2.7
packaging==24.2
paho-mqtt==2.1.0
sqlparse==0.5.3
tzdata==2024.2
